Move routes to separate module
